pyproject.toml
[project]
name = "registration-verification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.46",
    "flask>=3.1.2",
    "flask-mail>=0.10.0",
    "opencv-python-headless>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pymongo>=4.15.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "gspread>=6.1.4",
    "google-auth>=2.35.0",
    "mcp[cli]>=1.22.0",
    "httpx>=0.28.1",
    "langchain-openai>=1.1.0",
    "langchain-core>=1.1.0",
    "langgraph>=1.0.3",
]
